riscv,vspace: unmapPageTable clear correct entry
When unmapping a page table, this function searches through the page
table structures and finds a slot containing the page table we wish to
delete. Previously, the slot containing the page table wasn't being
correctly cleared as the slot pointer was updated to a slot in the page
table that was supposed to be deleted.  This led to behavior where
unmapping a page table would leave the table still mapped.