riscv: make consistent PageTable isMapped checks
A page table cap has at least the following entries:
- capPTMappedASID
- capPTIsMapped
- capPTMappedAddress

These are used to implement a type Option<(ASID, MappedAddress)> where
capPTIsMapped is used to reflect whether the values capPTMappedASID and
capPTMappedAddress are valid or not. Frame caps don't have an IsMapped
equivalent but instead check if capFMappedASID == 0.  When a Pagetable
is unmapped capPTIsMapped is cleared, but capPTMappedASID and
capPTMappedAddress are left unchanged.  When a frame is unmapped then
capFMappedASID will be set to 0.

It appears that the RISC-V port often uses
cap_page_table_cap_get_capPTMappedASID(cap) != asidInvalid to check
whether a page table cap is mapped. This can lead to incorrect behavior
if a page table has been mapped and then unmapped as the asid will be
left as a valid asid. Each of these changes are changed to instead
consistently check capPTIsMapped. In addition it is required that a when
a PageTable becomes reachable as either a vspace_root, or another entry
in the tree that capPTIsMapped is set.