RISC-V: Use 0 for UserError fault type
Using an address isn't valid as the field that it gets stored in has a
bitfield definition that isn't a word width. This fails an assert when
the value gets masked. For now we send 0, the same as on Arm. However,
the bitfield definition could be changed to allow reporting this info in
the future.