riscv: Prevent PageTableUnmap on top level PTs
It is only valid to call RISCVPageTableUnmap on page tables that are not
top level page tables. Deactivating a top level PageTable is done by
revoking the untyped that it was created from, which will invalidate
the ASID and all other mappings in the hierarchy.  This is consistent
with all other architectures.