exynos4, exynos5, sabrelite: clamp memory to fit in kernel window
This is a hack that allows user-space tools like the capDL static
allocator to work. Otherwise, the allocator (currently unaware of the
kernel window) would think that there is a 28-bit untyped at pptr
0xf0000000, and would generate untyped derivations that are not
available at runtime.

Note that we clamp to 0xff000000 on exynos4 and exynos5, *not*
0xfff00000 (PPTR_TOP), to avoid the ASID PD hole at
0xff000000..0xff200000. The affected platforms are ARM VM platforms;
VM images are often larger than 13MiB and would overlap the ASID hole
if loaded from PPTR_TOP. init_freemem fails if this overlap exists.