mcs: use head+length implementation refill queue
It had seemed previously that we could not used such a queue structure,
however we only ever access the head, head + 1, tail, and tail + 1
elements in the queue so we will only ever attempt to access one past
the end of the buffer which can easily be accounted for.

This allows us to have a consistent representation of an empty queue
which can allow for a number of simplifications in the refill logic.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>