mcs: Don't pass capacity through calls
The capacity does not need to be passed as an argument to
refill_check_budget as all information that it was being used for can be
dertermined from the usage directly.

The additional case handling for overrun into a partial refill is also
handled by split check, so that is used in all cases where any partial
usage remains.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>