exynos5: tweak CMake for greppability
Introduce a variable to hold a long expression to prevent the code
styler from line-wrapping the declare_platform() statement.  We want to
keep that on one line so the `griddle` tool (or humans) can easily
`grep` a list of supported platforms.  As of 2019-08-07, this platform
is the only one requiring this workaround.