aarch64, fpu: update hardware fpu support check
Cortex-A55 processors have additional support for half-precision
floating point arithmetic, which is defined as 0b0001 in
id_aa64pfr0_el1. This value includes all support for single and double
precision floating point types. This commit adds a macro checking the
register for 0b1111, which is universally defined as "Floating-point is
not implemented". All other values are reserved.

Co-authored-by: Alex Pavey <Alex.Pavey@dornerworks.com>

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>