aarch64, fpu: update hardware fpu support check
Some processors have additional support for half-precision
floating point arithmetic, which is defined as 0b0001 in
id_aa64pfr0_el1. This value includes all support for single and double
precision floating point types.

This commit adds a macro checking the register for 0b1111, which is
universally defined as "Floating-point is not implemented". FPU support
is then enabled if either the AdvSIMD or FP fields in id_aa64pfr0_el1
are != 0b1111 instead of requiring that both are not equal.

Co-authored-by: Alex Pavey <Alex.Pavey@dornerworks.com>

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>