mcs: ensure that all checks are performed on ticks
Although the configuration of WCET, domain time, and scheduling
contexts are provided in milliseconds or microseconds, all of the
comparisons must occur after the values have been converted to timer
ticks.

This is needed as the conversion is not guaranteed to preserve the
properties of the checks due to possible truncation.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>