mcs: check lower bounds in ticks
As conversion from us to ticks can be lossy, checking the minimum budget
on the us doesn't always guarantee that the time will still be greater
that the minimum after conversion.

Whilst the upper bounds on time are still checked in us to ensure that
overflow does not occur, the lower bounds checks are performed after
conversion to ensure that the lower bound isn't subverted by lossy
conversion.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>