armv8,CMake: Split generic timer, counter access
Instead of exporting the generic timer and counter access to userland in
one go, this commit splits it up into two different operations which
require two different build flags to be turned on. This is because the
timers expose a storage channel and allow threads to corrupt the
registers for each other.

Also move the config options to a better location and have them
depend on a more suitable config flag.