mcs: Sender can't control SC donation
Senders of IPCs to passive servers are not given any control as to
whether SCs are donated so as to avoid the situation where a passive
server expects an SC but ends up inactive with no SC due to misplaced
trust in the sender.

The new semantics for scheduling context donation are simply that if the
sender has not timeout faulted, the receiver has no SC, and the sender
has an SC then the sender's SC will be donated to the receiver.

If the sender has an SC it is guaranteed to be ready on its core unless
it is a time-out fault (in which case the SC won't be donated to the
receiver).

This ensures a receiver will never time-out fault but that it may become
unschedulable if it ends up without an SC. To prevent this, trusted
software should not remove scheduling contexts from TCBs blocked on send
to endpoints serviced by passive servers.