boot: Use extra_bi_size_bits for vreg size calc
The bootinfo extra size region is rounded up to a power of 2 size before
being mapped into the initial threads address space.  This wasn't being
handled correctly which lead to a situation where if the bootinfo
finishes on the last address of a page table but the extra region size
is not a power of 2 size then there won't be a page directory allocated
where one needs to be and the kernel would crash.