Kernel-Vtx: Use clang compatible inline assembly
Specifying a function name prefixed with 'm' is not compatible with
clang.

For 32 bit, call vmlaunch_failed explicitly from within the assembly.
Requires adding the USED attribute to vmlaunch_failed as it is only
referenced in a string literal.

For 64 bit, move the address of vmlaunch_failed into rax as an integer
rather than loading it as the address of a function (lea).