Don't unblock current SC
If the current SC is bound to a notification and the current thread is
deleted it is possible for the current SC to be signalled by the
pre-emption. In this case the current SC may be donated to another
thread without being changed the value in ksConsumed.

refill_unblock_check can only be called on the current SC if
ksConsumed is 0. Further, it makes no sense to call refill_unblock_check
on the current SC as it was not blocked.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>