libsel4,riscv: Remove seL4_UserException_FLAGS
This field is unused and is never set to anything by the kernel. Other
architecture definitions of seL4_UserException_Msg contain an equivalent
field for a current-program-status-register-like register that the fault
handler can read and modify as part of fault handling. RISC-V's
equivalent register SSTATUS doesn't contain any fields that would be
useful for a fault handler to modify at this stage and so the register
isn't transferred in the fault.

Signed-off-by: Kent McLeod <Kent.Mcleod@data61.csiro.au>