riscv: fix page mapping bug
Previously, it was possible in decodeRISCVFrameInvocation for the rwx
rights of the new PTE to become 000 after masking with cap rights. This
would turn the frame PTE into a page table PTE instead, and allow the
user to create almost arbitrary mappings, including to kernel data and
code. The defect was discovered in the C verification of the RISC-V
port.

This commit changes the PTE to an invalid PTE in that situation, which
has the same effect (faulting when accessed) as mapping a PTE without
rights has on the other supported architectures.

Signed-off-by: Gerwin Klein <gerwin.klein@data61.csiro.au>