riscv: change top-level page table unmap check
The basic model of seL4 system calls is decode+perform. To be a
refinement of higher-level models, we cannot in a decode decide that
everything is fine, but we will do nothing.

In this case, we need to ensure the page table being unmapped is not a
top-level page table, by looking up the ASID. A page table that isn't
mapped does not have a stored ASID. performPageTableInvocationUnmap is
capable of handling this and there is no need to early-exit from
decodeRISCVPageTableInvocation.

We proceed with performing an unmap on a non-mapped page table
regardless, and only check for being a top-level page table in the
mapped case.

Signed-off-by: Rafal Kolanski <rafal.kolanski@data61.csiro.au>