Repair barriers in try_arch_atomic_exchange
try_arch_atomic_exchange currently never acts as a release barrier.
The reason for this is that a release fence needs to be inserted before
the store it attempts to order with previous stores.
In the current implementation, a release barrier is only inserted after
the store, thus the exchange operation is never ordered w.r.t. previous
operations.
With the suggested fix, passing RELEASE or stronger as success memory
order will correctly order the exchange operation w.r.t.
previous operations.

Signed-off-by: jonas <s9joober@gmail.com>