mcs: All refills are at least MIN_BUDGET
All refills must always be at least MIN_BUDGET so we can determine that
any refill will be sufficient to run when it becomes the head.

By ensuring that the total budget of a scheduling context is at least
twice MIN_BUDGET, we can ensure that all refills are at least MIN_BUDGET
and that it is always possible to fragment into at least 2 refills.

This gives us that when performing sched_used it can either split the
tail refill in 2 parts (each of at least MIN_BUDGET) and combine the
second ahead of the new refill, or that there is already at least 2
refills. It then follows that sched_used should _never_ cause the head
refill to be moved.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>