mcs: Add check to fault thread if unschedulable
If we would reschedule a thread but we do not already know the state of
the SC associated with that thread we need to ensure that the thread has
an SC that is configured and has a released head.

If it has a timeout endpoint and it is nonexistant, not configured, or
not released, the appropriate timeout is sent to the timeout endpoint.

If it has no timeout endpoint and is not configured or released then it
is simply not scheduled. Otherwise, if the head refill is not released,
it is entered into the release queue for a later release.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>