mcs: Unfold schedContext_resume in refill_update
This is the only case where we would rescheudle a thread but do not
fault it if it is not schedulable. This is because refill_update
preserves the schedulability of the thread currently using the SC. It
will never make a configured SC unconfigured and it will never make a
ready SC unready.

We unfold the call to schedContext_resume to remove checks that are
statically guaranteed at that point in the code and reduce the number of
overall checks there, calling postpone if the SC is unready.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>