mcs: ensure schedulable when resuming thread
This replaces any occurence of schedContext_resume (which performed a
subset of the behavior of ensureSchedulable) and adds checks to any
location where a thread may be scheduled but we don't have any static
guarantees on it being schedulable.

The sendIPC case is handled in a separate commit to avoid recursion.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>